Move EmailRequestModel and EmailTemplate enum to Commmon
