+++ backend/src/StamAcasa.EmailService/EmailBuilder/IEmailBuilderService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using StamAcasa.Common.EmailService;
 using StamAcasa.Common.Models;
 
 namespace StamAcasa.EmailService.EmailBuilder
