+++ backend/src/StamAcasa.EmailService/EmailBuilder/ITemplateFileSelector.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using StamAcasa.Common.EmailService;
 
 namespace StamAcasa.EmailService.EmailBuilder
 {
     public interface ITemplateFileSelector
     {
-        string GetTemplatePath(Models.EmailTemplate template);
+        string GetTemplatePath(EmailTemplate template);
     }
 }
